requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
plotly
pandas
streamlit-option-menu